Refactor: Separate contextual parsing from Execer
